Use ssize_t instead of long in mtrand

+++ numpy/random/mtrand/initarray.h
@@ -2,5 +2,4 @@
 
 extern void
 init_by_array(rk_state *self, unsigned long init_key[], 
-              unsigned long key_length);
-
+              size_t key_length);
